requirements.txt
asgiref==3.6.0
awsebcli==3.20.10
backports.zoneinfo==0.2.1
blessed==1.20.0
boto3==1.26.134
botocore==1.29.134
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.3
dj-database-url==2.2.0
Django==4.2.1
django-active-link==0.1.8
django-storages==1.13.2
django-widget-tweaks==1.5.0
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.1
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.12.2
urllib3==1.26.15
wcwidth==0.1.9
whitenoise==6.6.0
